renames twind.config.ts to tailwind.config.ts
