+++ tests/cli_test.ts
@@ -209,7 +209,7 @@ Deno.test({
           { "type": "file", "name": "README.md" },
           { "type": "file", "name": "import_map.json" },
           { "type": "file", "name": "fresh.gen.ts" },
-          { "type": "file", "name": "twind.config.ts" },
+          { "type": "file", "name": "tailwind.config.ts" },
           {
             "type": "directory",
             "name": "components",
