+++ update.ts
@@ -105,7 +105,7 @@ import { start } from "$fresh/server.ts";
 import manifest from "./fresh.gen.ts";
 
 import twindPlugin from "$fresh/plugins/twind.ts";
-import twindConfig from "./twind.config.ts";
+import twindConfig from "./tailwind.config.ts";
 
 await start(manifest, { plugins: [twindPlugin(twindConfig)] });\n`;
   const MAIN_TS_PATH = join(resolvedDirectory, "main.ts");
@@ -118,7 +118,7 @@ await start(manifest, { plugins: [twindPlugin(twindConfig)] });\n`;
   } as Options;
   `;
   await Deno.writeTextFile(
-    join(resolvedDirectory, "twind.config.ts"),
+    join(resolvedDirectory, "tailwind.config.ts"),
     TWIND_CONFIG_TS,
   );
