+++ www/main.ts
@@ -8,6 +8,6 @@ import { start } from "$fresh/server.ts";
 import twindPlugin from "$fresh/plugins/twind.ts";
 
 import manifest from "./fresh.gen.ts";
-import twindConfig from "./twind.config.ts";
+import twindConfig from "./tailwind.config.ts";
 
 await start(manifest, { plugins: [twindPlugin(twindConfig)] });
